requirements.txt
flask==3.1.0
flask-login==0.6.3
PyMuPDF==1.25.3
mammoth==1.8.0
html2text==2024.2.26
python-dotenv==1.0.1
python-slugify==8.0.4
